page and error

+++ controllers/Room.js
@@ -2,6 +2,7 @@
 const db = require('../models/index');
 const fs = require('fs').promises;
 const path = require('path');
+const page = require('../middleware/page');
 class Room {
   constructor(
     name,
@@ -23,9 +24,14 @@ class Room {
     this.delete_status = delete_status;
   }
   static async index(req, res) {
-    const roomDb = await db.Room.findAll();
-    const rooms = roomDb.map(room => room.get({ plain: true }));
-    res.render('Room/index', { layout: 'admin', room: rooms });
+    const currentPage = req.query.page || 1;
+    const { objects, pagesArray } = await page(currentPage, db.Room);
+    res.render('Room/index', {
+      layout: 'admin',
+      room: objects,
+      pagesArray,
+      currentPage,
+    });
   }
   static async room(req, res) {
     const roomDb = await db.Room.findAll();
@@ -56,7 +62,7 @@ class Room {
     const room = new Room(name, type, floor, price, capacity, image_url);
     try {
       await db.Room.create(room);
-      res.redirect('/room');
+      res.redirect('/room/roomIndex');
     } catch (err) {
       console.log(err);
     }
@@ -76,18 +82,22 @@ class Room {
   }
 
   static async update(req, res) {
+    let image_url;
     const id = req.params.id;
-    const image_url = req.file.filename;
     const { name, type, floor, price, capacity, oldimage } = req.body;
-    const oldPath = path.join('views/images/', oldimage);
-    console.log(image_url);
-    await fs.unlink(oldPath);
+    if (req.file?.filename) {
+      image_url = req.file.filename;
+      const oldPath = path.join('views/images/', oldimage);
+      await fs.unlink(oldPath);
+    } else {
+      image_url = oldimage;
+    }
     try {
       await db.Room.update(
         { name, type, floor, price, capacity, image_url },
         { where: { id: id } }
       );
-      res.redirect('/room');
+      res.redirect('/room/roomIndex');
     } catch (err) {
       console.log(err);
     }
@@ -101,6 +111,16 @@ class Room {
       res.redirect('/room');
     } catch (err) {
       console.log(err);
+    }
+  }
+
+  static async checkExist(req, res) {
+    const { name } = req.body;
+    const room = await db.Room.findOne({ where: { name: name } });
+    if (room) {
+      res.json({ exist: true });
+    } else {
+      res.json({ exist: false });
     }
   }
 }
